nodes: add ColorStats node
This node computes color statistics (color distribution essentially) of
a given texture, and expose the information a block. The following
commits will introduce the interface with other nodes.

From an interface point of view, it has the particularity of filling the
block data in the draw callback instead of the update: this is because
the texture could be changed after ColorStats.update() is called. This
typically happens if a RTT or another compute writes into it.