build: bump pkgconf from 1.9.4 to 1.9.5
We also change the source URL because the official distribution is down
and doesn't seem it's going to come up again anytime soon.

The original source was preferred due to its more convenient tree layout
and its inclusion of generated autotools files. But the fancy tree
layout is handled fine by our build system, and we use meson instead of
the autotools now, so these are not issues anymore.