glsl/path: make outline twice smaller
To understand why this is a better scale, let's look at the stroke width
calculation in the shader before this change:
    float stroke_d = outline_width - abs(dist.y)

Let's give x the distance and w the outline_width user parameter.

|x| is a pyramidal "/\" around x=0. The stroke width is w-|x|. When w=1,
the pyramid starts at -1 and ends at 1, which is an amplitude of 2,
which means the stroke is twice as large as indicated by the user input.